use of double in mandatory instead of float

+++ srcs/init_matrix.c
@@ -12,25 +12,25 @@
 
 #include "fdf.h"
 
-/// @brief Allocate memory for an array of array of floats.
+/// @brief Allocate memory for an array of array of doubles.
 /// @return Return an array of array of pointer (matrix[3][3]).
-float	**init_matrix(void)
+double	**init_matrix(void)
 {
-	float	**matrix;
+	double	**matrix;
 
-	matrix = malloc(3 * sizeof(float *));
+	matrix = malloc(3 * sizeof(double *));
 	if (!matrix)
 		return (NULL);
-	matrix[0] = malloc(3 * sizeof(float));
+	matrix[0] = malloc(3 * sizeof(double));
 	if (!matrix[0])
 		return (free(matrix), NULL);
-	matrix[1] = malloc(3 * sizeof(float));
+	matrix[1] = malloc(3 * sizeof(double));
 	if (!matrix[1])
 	{
 		free(matrix[0]);
 		return (free(matrix), NULL);
 	}
-	matrix[2] = malloc(3 * sizeof(float));
+	matrix[2] = malloc(3 * sizeof(double));
 	if (!matrix[2])
 	{
 		free(matrix[0]);
@@ -40,11 +40,11 @@ float	**init_matrix(void)
 	return (matrix);
 }
 
-/// @brief Allocate memory for an array of array of floats used in X rotation.
+/// @brief Allocate memory for an array of array of doubles used in X rotation.
 /// @return Return an array of array of pointer (matrix[3][3]).
-float	**init_matrix_x(t_data data)
+double	**init_matrix_x(t_data data)
 {
-	float	**matrix;
+	double	**matrix;
 
 	matrix = init_matrix();
 	if (!matrix)
@@ -61,11 +61,11 @@ float	**init_matrix_x(t_data data)
 	return (matrix);
 }
 
-/// @brief Allocate memory for an array of array of floats used in Y rotation.
+/// @brief Allocate memory for an array of array of doubles used in Y rotation.
 /// @return Return an array of array of pointer (matrix[3][3]).
-float	**init_matrix_y(t_data data)
+double	**init_matrix_y(t_data data)
 {
-	float	**matrix;
+	double	**matrix;
 
 	matrix = init_matrix();
 	if (!matrix)
@@ -82,11 +82,11 @@ float	**init_matrix_y(t_data data)
 	return (matrix);
 }
 
-/// @brief Allocate memory for an array of array of floats used in Z rotation.
+/// @brief Allocate memory for an array of array of doubles used in Z rotation.
 /// @return Return an array of array of pointer (matrix[3][3]).
-float	**init_matrix_z(t_data data)
+double	**init_matrix_z(t_data data)
 {
-	float	**matrix;
+	double	**matrix;
 
 	matrix = init_matrix();
 	if (!matrix)
